telegram-bot/requirements.txt
python-telegram-bot==20.8
requests
python-dotenv
flask
gunicorn
